requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
mysql-connector-python==8.4.0
bcrypt==4.0.1